requirements.txt
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
black>=23.7.0
pylint>=2.17.5
xvfbwrapper>=0.2.9
pytest-xvfb>=2.0.0
typing-extensions>=4.7.1  # For Python type hints
pathlib>=1.0.1  # For path handling